feat: create catalogue filters component
