+++ src/components/CatalogueFilters/CatalogueFilters.scss
@@ -1,0 +1,40 @@
+.filters {
+  display: flex;
+  flex-direction: column;
+  background-color: white;
+  padding: 10px;
+  gap: 20px;
+
+  &__checkbox_brand {
+    height: 200px;
+    overflow: hidden scroll;
+    scrollbar-color: #333333 #f9f9f9;
+    scrollbar-width: thin;
+  }
+
+  &__copy-link {
+    margin-top: 30px;
+    width: 100%;
+    height: 50px;
+    background-color: #333333;
+    color: #f9f9f9;
+  }
+
+  &__copy-link:hover {
+    background-color: #f9f9f9;
+    color: #333333;
+    border: 1px solid #333333;
+    transition: 0.5s;
+  }
+}
+
+::-webkit-scrollbar {
+  width: 5px;
+  background-color: #f9f9f9;
+}
+
+::-webkit-scrollbar-thumb {
+  background-color: #333333;
+  border-radius: 9em;
+  box-shadow: inset 1px 1px 10px #333333;
+}
