+++ src/pages/CataloguePage/CataloguePage.tsx
@@ -2,16 +2,11 @@ import React from "react";
 import data from "../../assets/data/data.json";
 import Button from "../../components/Buttons/Button/Button";
 import CatalogueSettings from "../../components/CatalogueSettings/CatalogueSettings";
-import RangeInput from "../../components/RangeInput/RangeInput";
-import CheckboxInput from "../../components/CheckboxInput/CheckboxInput";
-import { WineType, Titles } from "../../utils/helpers/interfaces";
 import heartIcon from "../../assets/img/svg/heart-icon.svg";
 import cartIcon from "../../assets/img/svg/cart-icon.svg";
 import searchIcon from "../../assets/img/svg/search-icon.svg";
 import "./CataloguePage.scss";
-
-const types = Object.values(WineType);
-const titles = Object.values(Titles);
+import CatalogueFilters from "../../components/CatalogueFilters/CatalogueFilters";
 
 const CataloguePage: React.FC = () => {
   return (
@@ -19,23 +14,7 @@ const CataloguePage: React.FC = () => {
       <div className="container">
         <CatalogueSettings />
         <div className="filters-goods-wrapper">
-          <div className="filters">
-            <CheckboxInput legend="TYPE OF WINE" optionsArray={types} />
-            <CheckboxInput
-              legend="BRAND"
-              optionsArray={titles}
-              className="filters__checkbox_brand"
-            />
-            <RangeInput
-              title="PRICE"
-              id="price"
-              minVal="$5.99"
-              maxVal="$56"
-              className="price-range"
-            />
-            <RangeInput title="QUANTITY" id="quantity" minVal="4" maxVal="22" />
-            <Button className="filters__copy-link">Copy Link</Button>
-          </div>
+          <CatalogueFilters />
           <div className="goods">
             {data.goods.map((elem) => (
               <div className="goods__item">
